split types and constants from main file
