+++ mlol_client/__init__.py
@@ -1,1 +1,2 @@
-from .mlol_client import MLOLClient, MLOLBook, MLOLLoan, MLOLReservation
+from .mlol_client import MLOLClient
+from .mlol_types import MLOLBook, MLOLLoan, MLOLReservation, MLOLUser
