+++ mlol_client/mlol_client.py
@@ -17,161 +17,14 @@ from requests.models import Response
 from requests.packages.urllib3.util.retry import Retry
 from requests_toolbelt import sessions
 
-LIBRARY_MAPPING_FNAME = os.path.join(os.path.dirname(__file__), "library_mapping.json")
-
-DEFAULT_USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/87.0.4280.67 Safari/537.36"
-DEFAULT_WEB_HEADERS = {
-    "User-Agent": DEFAULT_USER_AGENT,
-    "Upgrade-Insecure-Requests": "1",
-    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
-    "Sec-Fetch-Site": "same-origin",
-    "Sec-Fetch-Mode": "navigate",
-    "Sec-Fetch-User": "?1",
-    "Sec-Fetch-Dest": "document",
-}
-
-WEB_ENDPOINTS = {
-    "index": "/home/index.aspx",
-    "search": "/media/ricerca.aspx",
-    "login": "/user/login.aspx",
-    "resources": "/user/risorse.aspx",
-    "get_book": "/media/scheda.aspx",
-    "redownload": "/help/dlrepeat.aspx",
-    "download": "/media/downloadebadok.aspx",
-    "pre_reserve": "/media/prenota.aspx",
-    "reserve": "/media/prenota2.aspx",
-    "cancel_reservation": "/media/annullaPr.aspx",
-    "get_queue_position": "/commons/QueuePos.aspx",
-}
-
-# taken from MLOL mobile app
-DEFAULT_API_HEADERS = {
-    "Host": "api.medialibrary.it",
-    "Connection": "Keep-Alive",
-    "Accept-Encoding": "gzip",
-    "User-Agent": "okhttp/3.9.0",
-}
-
-API_ENDPOINTS = {
-    "login": "https://api.medialibrary.it/app/login",
-    "portals": "https://api.medialibrary.it/app/portals",
-    "loan_history": "https://api.medialibrary.it/app/loanhistory",
-    "loans": "https://api.medialibrary.it/app/loans",
-    "userinfo": "https://api.medialibrary.it/app/profile",
-}
-
-
-class MLOLUser:
-    def __init__(
-        self,
-        id: int,
-        name: str,
-        surname: str,
-        username: str,
-        remaining_loans: int,
-        remaining_reservations: int,
-        expiration_date: datetime,
-    ):
-        self.id = id
-        self.name = name
-        self.surname = surname
-        self.username = username
-        self.remaining_loans = remaining_loans
-        self.remaining_reservations = remaining_reservations
-        self.expiration_date = expiration_date
-
-    def __repr__(self):
-        values = {
-            k: "{}{}".format(str(v)[:50], "..." if len(str(v)) > 50 else "")
-            for k, v in self.__dict__.items()
-            if v is not None
-        }
-        return f"<mlol_client.MLOLUser: {values}>"
-
-
-class MLOLBook:
-    def __init__(
-        self,
-        *,
-        id: str,
-        title: str,
-        authors: List[str] = None,
-        status: str = None,
-        publisher: str = None,
-        ISBNs: List[str] = None,
-        language: str = None,
-        description: str = None,
-        year: int = None,
-        formats: List[str] = None,
-        drm: bool = None,
-    ):
-        self.id = str(id)
-        self.title = title
-        self.authors = authors
-        self.status = status
-        self.publisher = publisher
-        self.ISBNs = ISBNs
-        self.language = language
-        self.description = description
-        self.year = year
-        self.formats = formats
-        self.drm = drm
-
-    def __repr__(self):
-        values = {
-            k: "{}{}".format(str(v)[:50], "..." if len(str(v)) > 50 else "")
-            for k, v in self.__dict__.items()
-            if v is not None
-        }
-        return f"<mlol_client.MLOLBook: {values}>"
-
-
-class MLOLReservation:
-    def __init__(
-        self,
-        *,
-        id: str,
-        book: MLOLBook,
-        date: datetime = None,
-        status: str = None,
-        queue_position: int = None,
-    ):
-        self.id = str(id)
-        self.book = book
-        self.date = date
-        self.status = status
-        self.queue_position = queue_position
-
-    def __repr__(self):
-        values = {
-            k: "{}{}".format(str(v)[:50], "..." if len(str(v)) > 50 else "")
-            for k, v in self.__dict__.items()
-            if v is not None
-        }
-        return f"<mlol_client.MLOLReservation: {values}>"
-
-
-class MLOLLoan:
-    def __init__(
-        self,
-        *,
-        id: str,
-        book: MLOLBook,
-        start_date: datetime = None,
-        end_date: datetime = None,
-    ):
-        self.id = str(id) if id else None
-        self.book = book
-        self.start_date = start_date
-        self.end_date = end_date
-
-    def __repr__(self):
-        values = {
-            k: "{}{}".format(str(v)[:50], "..." if len(str(v)) > 50 else "")
-            for k, v in self.__dict__.items()
-            if v is not None
-        }
-        return f"<mlol_client.MLOLLoan: {values}>"
+from .mlol_constants import (
+    WEB_ENDPOINTS,
+    API_ENDPOINTS,
+    DEFAULT_API_HEADERS,
+    DEFAULT_WEB_HEADERS,
+    LIBRARY_MAPPING_FNAME,
+)
+from .mlol_types import MLOLBook, MLOLLoan, MLOLReservation, MLOLUser
 
 
 class MLOLApiConverter:
